pyproject.toml
[project]
name = "multimedia-bot"
version = "0.1.0"
description = "Telegram inline media bot MVP on aiogram 3"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.20.0,<4.0.0",
    "aiosqlite>=0.21.0,<0.22.0",
    "alembic>=1.16.0,<2.0.0",
    "pillow>=11.2.0,<12.0.0",
    "pydantic-settings>=2.9.0,<3.0.0",
    "sqlalchemy>=2.0.40,<3.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5,<9.0.0",
    "pytest-asyncio>=0.26.0,<0.27.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.uv]
default-groups = ["dev"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/multimedia_bot"]
